+++ internal/server/bin.go
@@ -528,9 +528,12 @@ func (h *hub) ejectFromBin(players map[int32]*tracked, pos simPos, state uint32)
 				yaw = 270
 			}
 			sd := &armorStand{eid: h.allocEID(), dim: pos.dim,
-				x: float64(front.x) + 0.5, y: float64(front.y), z: float64(front.z) + 0.5, yaw: yaw}
+				x: float64(front.x) + 0.5, y: float64(front.y), z: float64(front.z) + 0.5, yaw: yaw, name: st.name}
 			h.armorStands[sd.eid] = sd
 			h.toNearbyEv(players, sd.dim, sd.x, sd.z, h.standAddEv(sd))
+			if sd.name != "" {
+				h.toNearbyEv(players, sd.dim, sd.x, sd.z, metaEv(nameMeta(sd.eid, sd.name)))
+			}
 			h.rsSound(players, "minecraft:entity.armor_stand.place", sndBlock, sd.x, sd.y, sd.z, 0.75, 0.8)
 		} else if it := h.spawnItemIn(players, h.rsDim, item, 1, fx, fy, fz); it != nil {
 			it.dmg, it.ench = st.dmg, st.ench